Renamed FuriousSlash to FuriousSlashUptime
