+++ src/Parser/Warrior/Fury/CombatLogParser.js
@@ -13,8 +13,8 @@ import Juggernaut from './Modules/BuffDebuff/Juggernaut';
 import RampageFrothingBerserker from './Modules/Features/RampageFrothingBerserker';
 import RampageCancelled from './Modules/Features/RampageCancelled';
 import AngerManagement from './Modules/Talents/AngerManagement';
-import FuriousSlash from './Modules/Talents/FuriousSlash';
 import FuriousSlashTimesByStacks from './Modules/Talents/FuriousSlashTimesByStacks';
+import FuriousSlashUptime from './Modules/Talents/FuriousSlashUptime';
 
 import T21_2set from './Modules/Items/T21_2set';
 import T21_4set from './Modules/Items/T21_4set';
@@ -37,8 +37,8 @@ class CombatLogParser extends CoreCombatLogParser {
     rampageFrothingBerserker: RampageFrothingBerserker,
     rampageCancelled: RampageCancelled,
     angerManagement: AngerManagement,
-	furiousSlash: FuriousSlash,
 	furiousSlashTimesByStacks: FuriousSlashTimesByStacks,
+	furiousSlashUptime: FuriousSlashUptime,
 
     t21_2set: T21_2set,
     t21_4set: T21_4set,
